@@ -43,7 +43,7 @@
     "cross-spawn": "^5.0.1",
     "date-fns": "^2.28.0",
     "echarts": "^5.3.2",
-    "eslint": "^8.16.0",
+    "eslint": "^8.18.0",
     "eslint-plugin-html": "^6.2.0",
     "highlight.js": "^11.5.1",
     "html-escaper": "^3.0.3",
