@@ -1,5 +1,5 @@
 <script setup>
-import { mapState, storeToRefs } from 'pinia'
+import { storeToRefs } from 'pinia'
 import { useI18n } from 'vue-i18n'
 import { purify } from '@/util/helper'
 import { useUserStore } from '@/store/modules/user'
@@ -11,36 +11,20 @@ const uid = $ref('')
 const newPwd = $ref('')
 const checkPwd = $ref('')
 
-// export default {
-//   data: () => ({
-//     uid: '',
-//     newPwd: '',
-//     checkPwd: '',
-//   }),
-//   computed: {
-//     ...mapState(useUserStore, [ 'user' ]),
-//   },
-//   created () {
-//   },
-//   methods: {
-//     findUser () {
-//       useUserStore().findOne({ uid: this.uid })
-//     },
-//     saveUser () {
-//       if (this.newPwd === this.checkPwd) {
-//         const user = purify(Object.assign(
-//           this.user,
-//           { newPwd: this.newPwd },
-//         ))
-//         useUserStore().update(user).then(() => {
-//           this.$Message.success('修改成功！')
-//         })
-//       } else {
-//         this.$Message.info('两次密码不一致，请重新输入！')
-//       }
-//     },
-//   },
-// }
+const findUser = () => userStore.findOne({ uid })
+
+async function saveUser () {
+  if (newPwd === checkPwd) {
+    const newUser = purify(Object.assign(
+      user,
+      { newPwd: this.newPwd },
+    ))
+    await userStore.update(newUser)
+    Message.success(t('oj.update_success'))
+  } else {
+    Message.info(t('oj.password_not_match'))
+  }
+}
 </script>
 
 <template>
